Fixed most issues from the CR
